修改: 節能查核成效資料流

+++ Taipei-City-Dashboard-DE/dags/proj_city_dashboard/energy_saving_coaching_situation_tpe_gov/energy_saving_coaching_situation_tpe_gov.py
@@ -65,7 +65,7 @@ def _transfer(**kwargs):
     data['compliance_rate'] = data['compliance_rate'].replace('%', '', regex=True)
     data['compliance_rate'] = data['compliance_rate'].astype(float)
     data['reinspection_compliance_rate'] = data['reinspection_compliance_rate'].replace('%', '', regex=True)
-    data['reinspection_compliance_rate'] = data['reinspection_compliance_rate'].astype(float)
+    # data['reinspection_compliance_rate'] = data['reinspection_compliance_rate'].astype(float)
     data['audited_households'] = data['audited_households'].astype(int)
     data['carbon_sink_daan_forest_parks'] = data['carbon_sink_daan_forest_parks'].astype(int)
     data['power_saving_ten_thousand_kWh_per_year'] = data['power_saving_ten_thousand_kWh_per_year'].astype(int)
